+++ src/main/java/com/safety/safetynet/repository/AllergiesRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AllergiesRepository extends JpaRepository<Allergies, Long> {
+public interface AllergiesRepository extends JpaRepository<Allergies, Integer> {
 }
